finished SkipQueue insertion algorithm

+++ processGraphs.c
@@ -138,7 +138,7 @@ int *shortestPath(int source, Graph *g){
     return dist;
 }
 
-int ASPL(Graph *g){
+float ASPL(Graph *g){
     double ASPL = 0;
     for (int i = 0; i < g->num_vertices; i++){
         double *distances = shortest_path(i, g);
@@ -146,5 +146,5 @@ int ASPL(Graph *g){
             ASPL += (distances[j] / 2.0);
         }
     }
-    return (int)ASPL;
+    return ASPL;
 }
